fix(js): check if asset create/update event contains a file before copying (#12551)
(cherry picked from commit 136e1a81fe5e4fd85fbea91d3b8a28d003987df8)

+++ packages/js/src/utils/assets/copy-assets-handler.ts
@@ -36,7 +36,9 @@ export const defaultFileEventHandler = (events: FileEvent[]) => {
   dirs.forEach((d) => fse.ensureDirSync(d));
   events.forEach((event) => {
     if (event.type === 'create' || event.type === 'update') {
-      fse.copyFileSync(event.src, event.dest);
+      if (fse.lstatSync(event.src).isFile()) {
+        fse.copyFileSync(event.src, event.dest);
+      }
     } else if (event.type === 'delete') {
       fse.removeSync(event.dest);
     } else {
